chore(vth-dashboard): update the config-sync json
After upgrading Strapi, new UI and API configurations have been added to the dashboard.

Please note that the config-sync plugin currently has a known bug, which may cause it to not function correctly. The issue is documented here:
[config-sync plugin issue #110](https://github.com/boazpoolman/strapi-plugin-config-sync/issues/110).

Although the issue has been fixed by the developers, there was a problem with the unit test on the GitHub action. Once the patch version is released, the issue will be resolved.
In the meantime, you can test the updated version locally by installing the package from GitHub using the following method:

```json
    "strapi-plugin-config-sync": "github:boazpoolman/strapi-plugin-config-sync",
``